feat: Changed quantities to decimal

+++ Application/Core/AutoMapperProfile.cs
@@ -66,7 +66,7 @@ namespace Application.Core
             CreateMap<Purchase, PurchaseCreateParam>().ReverseMap();
             CreateMap<Purchase, PurchaseDto>()
                 .ForMember(pd => pd.ProductDescription, opt => opt.MapFrom(p => p.Product.ProductName))
-                .ForMember(pd => pd.QuantityRemaining, opt => opt.MapFrom(p => (decimal)p.Product.Quantity))
+                .ForMember(pd => pd.QuantityRemaining, opt => opt.MapFrom(p => p.Product.Quantity))
                 .ForMember(pd => pd.ProductDefaultImage, opt => opt.MapFrom(p => p.Product.ProductPhotos.Aggregate("", (url, next)=> next.Url != null? next.Url: "" )))
                 .ForMember(pd => pd.PurchaseState, opt => opt.MapFrom(p => p.Order.OrderState))
                 .ForMember(
